+++ visualization/app/codeCharta/state/settings.service.ts
@@ -1,5 +1,6 @@
 import {
-	ColorRange, DynamicSettings,
+	ColorRange,
+	DynamicSettings,
 	FileSettings,
 	FileState,
 	MapColors,
@@ -7,11 +8,12 @@ import {
 	Settings
 } from "../codeCharta.model"
 import _ from "lodash"
-import { IAngularEvent, IRootScopeService } from "angular"
+import {IAngularEvent, IRootScopeService, ITimeoutService} from "angular"
 import {FileStateService, FileStateServiceSubscriber} from "./fileState.service";
 import {FileStateHelper} from "../util/fileStateHelper";
 import {SettingsMerger} from "../util/settingsMerger";
 import {Vector3} from "three"
+import {RibbonBarController} from "../ui/ribbonBar/ribbonBar.component";
 
 export interface SettingsServiceSubscriber {
 	onSettingsChanged(settings: Settings, update : RecursivePartial<Settings>, event: IAngularEvent)
@@ -22,11 +24,14 @@ export class SettingsService implements FileStateServiceSubscriber {
 	private static SETTINGS_CHANGED_EVENT = "settings-changed"
 
 	private settings: Settings
-	private readonly throttledBroadcast: (update : RecursivePartial<Settings>) => void
+	private readonly debounceBroadcast: (update : RecursivePartial<Settings>) => void
 
-	constructor(private $rootScope) {
+	constructor(
+		private $rootScope: IRootScopeService,
+		private $timeout: ITimeoutService
+	) {
 		this.settings = this.getDefaultSettings()
-		this.throttledBroadcast = _.throttle((update : RecursivePartial<Settings>) =>
+		this.debounceBroadcast = _.debounce((update : RecursivePartial<Settings>) =>
 			this.$rootScope.$broadcast(SettingsService.SETTINGS_CHANGED_EVENT, { settings: this.settings, update: update}), 400)
 		FileStateService.subscribe(this.$rootScope, this)
 	}
@@ -51,8 +56,10 @@ export class SettingsService implements FileStateServiceSubscriber {
 		// _.merge(this.settings, update) didnt work with arrays like blacklist
 		this.settings = this.updateSettingsUsingPartialSettings(this.settings, update)
 		if(!isSilent) {
-			this.throttledBroadcast(update)
+			this.$rootScope.$broadcast(RibbonBarController.LOADING_MAP_STATUS_EVENT, true)
+			this.debounceBroadcast(update)
 		}
+		this.synchronizeAngularTwoWayBinding()
 	}
 
 	public getDefaultSettings(): Settings {
@@ -161,6 +168,10 @@ export class SettingsService implements FileStateServiceSubscriber {
 			}
 		}
 		return false
+	}
+
+	private synchronizeAngularTwoWayBinding() {
+		this.$timeout(() => {})
 	}
 
 	public static subscribe($rootScope: IRootScopeService, subscriber: SettingsServiceSubscriber) {