+++ CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased]
+
 ### Added
 
+-   Show loading-gif in ribbonBar when rerendering map
+
 ### Changed
+
+-   Use debounced settings update instead of throttled
 
 ### Removed
 
@@ -18,14 +23,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Chore
 
 ## [1.24.0] - 2019-04-23
+
 ### Added
 
 ### Changed
 
 ### Removed
+
 -   Settings as URL parameters #470
 
 ### Fixed
+
 -   Fixed issue with trailing slash in URL parameter of SonarImporter #356
 
 ### Chore